// src/Types/Playlist/TracksTypes.ts
    export interface ExternalUrls {
        spotify: string;
    }

    export interface AddedBy {
        external_urls: ExternalUrls;
        href: string;
        id: string;
        type: string;
        uri: string;
    }

    export interface ExternalUrls2 {
        spotify: string;
    }

    export interface Artist {
        external_urls: ExternalUrls2;
        href: string;
        id: string;
        name: string;
        type: string;
        uri: string;
    }

    export interface ExternalUrls3 {
        spotify: string;
    }

    export interface Image {
        height: number;
        url: string;
        width: number;
    }

    export interface Album {
        album_type: string;
        artists: Artist[];
        available_markets: string[];
        external_urls: ExternalUrls3;
        href: string;
        id: string;
        images: Image[];
        name: string;
        release_date: string;
        release_date_precision: string;
        total_tracks: number;
        type: string;
        uri: string;
    }

    export interface ExternalUrls4 {
        spotify: string;
    }

    export interface Artist2 {
        external_urls: ExternalUrls4;
        href: string;
        id: string;
        name: string;
        type: string;
        uri: string;
    }

    export interface ExternalIds {
        isrc: string;
    }

    export interface ExternalUrls5 {
        spotify: string;
    }

    export interface Track {
        album: Album;
        artists: Artist2[];
        available_markets: string[];
        disc_number: number;
        duration_ms: number;
        episode: boolean;
        explicit: boolean;
        external_ids: ExternalIds;
        external_urls: ExternalUrls5;
        href: string;
        id: string;
        is_local: boolean;
        name: string;
        popularity: number;
        preview_url: string;
        track: boolean;
        track_number: number;
        type: string;
        uri: string;
    }

    export interface VideoThumbnail {
        url?: any;
    }

    export interface TracksTypes {
        added_at: Date;
        added_by: AddedBy;
        is_local: boolean;
        primary_color?: any;
        track: Track;
        video_thumbnail: VideoThumbnail;
    }